requirements.txt
numpy>=1.26.0
chromadb>=0.5.3
pandas>=2.0.0
pypdf>=5.1.0
openai>=1.60.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.29
langchain-text-splitters>=0.0.1
umap-learn>=0.5.0
matplotlib>=3.9.0
sentence-transformers>=2.0.0
transformers>=4.0.0
torch>=2.0.0
fastapi~=0.116.1
boto3~=1.40.1
pydantic~=2.11.7
starlette~=0.47.2
httpx~=0.28.1
spacy~=3.8.7